NeMo Data Designer/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pandas
nemo-microservices[data-designer]
streamlit